Refactoring styling
